+++ data/resources/CHANGELOG.md
@@ -1,3 +1,7 @@
+Jan 15, 2025:
+- live: Update for patch
+- lua: Add error handling to mq.delay, failures inside the callback should now propagate to the caller properly
+
 Jan 12, 2025:
 - test: Update for patch
 - Added ${Pet.ProcHold} and ${Pet.Resume}